Move internal API such as file_get_contents back into scout_internal_handlers
