+++ scout_execute_ex.c
@@ -7,12 +7,8 @@
 
 #include "zend_scoutapm.h"
 #include "scout_extern.h"
-#include "scout_execute_ex.h"
 
-#if SCOUTAPM_INSTRUMENT_USING_OBSERVER_API == 1
-extern zend_observer_fcall_handlers scout_observer_api_register(zend_execute_data *execute_data);
-#else
-void add_function_to_instrumentation(const char *function_name);
+#if SCOUTAPM_INSTRUMENT_USING_OBSERVER_API == 0
 void (*original_zend_execute_ex) (zend_execute_data *execute_data);
 void (*original_zend_execute_internal) (zend_execute_data *execute_data, zval *return_value);
 void scoutapm_execute_internal(zend_execute_data *execute_data, zval *return_value);
@@ -22,10 +18,6 @@ void scoutapm_execute_ex(zend_execute_data *execute_data);
 int setup_functions_for_zend_execute_ex()
 {
 #if SCOUTAPM_INSTRUMENT_USING_OBSERVER_API == 0
-    ADD_FUNCTION_TO_INSTRUMENTATION_SAFE_CATCH("file_get_contents");
-    ADD_FUNCTION_TO_INSTRUMENTATION_SAFE_CATCH("file_put_contents");
-    ADD_FUNCTION_TO_INSTRUMENTATION_SAFE_CATCH("pdo->exec");
-    ADD_FUNCTION_TO_INSTRUMENTATION_SAFE_CATCH("pdo->query");
     ADD_FUNCTION_TO_INSTRUMENTATION_SAFE_CATCH("Predis\\Client->get");
     ADD_FUNCTION_TO_INSTRUMENTATION_SAFE_CATCH("Predis\\Client->set");
     ADD_FUNCTION_TO_INSTRUMENTATION_SAFE_CATCH("Predis\\Client->del");
@@ -59,28 +51,6 @@ void deregister_scout_execute_ex()
 }
 
 #if SCOUTAPM_INSTRUMENT_USING_OBSERVER_API == 0
-
-void add_function_to_instrumentation(const char *function_name)
-{
-    if (SCOUTAPM_G(num_instrumented_functions) >= MAX_INSTRUMENTED_FUNCTIONS) {
-        zend_throw_exception_ex(NULL, 0, "Unable to add instrumentation for function '%s' - MAX_INSTRUMENTED_FUNCTIONS of %d reached", function_name, MAX_INSTRUMENTED_FUNCTIONS);
-        return;
-    }
-
-    SCOUTAPM_G(instrumented_function_names)[SCOUTAPM_G(num_instrumented_functions)] = strdup(function_name);
-    SCOUTAPM_G(num_instrumented_functions)++;
-}
-
-int should_be_instrumented(const char *function_name)
-{
-    int i = 0;
-    for (; i < SCOUTAPM_G(num_instrumented_functions); i++) {
-        if (strcasecmp(function_name, SCOUTAPM_G(instrumented_function_names)[i]) == 0) {
-            return 1;
-        }
-    }
-    return 0;
-}
 
 void scoutapm_execute_internal(zend_execute_data *execute_data, zval *return_value)
 {
