+++ scout_recording.h
@@ -10,8 +10,6 @@
 
 #include <zend_types.h>
 
-#define MAX_INSTRUMENTED_FUNCTIONS 100
-
 /* Describes information we store about a recorded stack frame */
 typedef struct _scoutapm_stack_frame {
     const char *function_name;
@@ -26,5 +24,13 @@ typedef struct _scoutapm_disconnected_call_argument_store {
     int argc;
     zval *argv;
 } scoutapm_disconnected_call_argument_store;
+
+#define ADD_FUNCTION_TO_INSTRUMENTATION_SAFE_CATCH(function_name)                                             \
+    zend_try {                                                                                                \
+        add_function_to_instrumentation(function_name);                                                       \
+    } zend_catch {                                                                                            \
+        php_printf("ScoutAPM tried instrumenting '%s' - increase MAX_INSTRUMENTED_FUNCTIONS", function_name); \
+        return FAILURE;                                                                                       \
+    } zend_end_try()
 
 #endif //SCOUTAPM_SCOUT_RECORDING_H
